feat(AngleWidget): Refactor and watch manipulator
