+++ src/vtk/Distance2DWidget/behavior.js
@@ -1,31 +1,8 @@
 import superWidgetBehavior from 'vtk.js/Sources/Widgets/Widgets3D/DistanceWidget/behavior';
 
+import widgetManipulatorWatcher from 'paraview-glance/src/vtk/widgetManipulatorWatcher';
+
 export default function widgetBehavior(publicAPI, model) {
   superWidgetBehavior(publicAPI, model);
-  const superBehavior = { ...publicAPI };
-
-  const watchers = [];
-  let prevCallData = null;
-
-  function retriggerMouseMove() {
-    if (prevCallData) {
-      superBehavior.handleMouseMove(prevCallData);
-    }
-  }
-
-  publicAPI.handleMouseMove = (callData) => {
-    const ret = superBehavior.handleMouseMove(callData);
-    prevCallData = callData;
-    return ret;
-  };
-
-  publicAPI.delete = () => {
-    superBehavior.delete();
-    while (watchers.length) {
-      watchers.pop().unsubscribe();
-    }
-  };
-
-  // re-place the point when manipulator changes
-  watchers.push(model.manipulator.onModified(() => retriggerMouseMove()));
+  widgetManipulatorWatcher(publicAPI, model);
 }
